refactor(rsc): organize plugin utils
