+++ packages/plugin-rsc/src/plugins/vite-utils.ts
